Refactoring derivation methods into PublicKey class. Moved Ed25519Extensions into Solnet.Wallet.
